Create PayoutDetailsAddress subcomponent and edit address config
