+++ src/translations/en.json
@@ -480,6 +480,9 @@
   "PayoutDetailsForm.postalCodeLabel": "Postal code",
   "PayoutDetailsForm.postalCodePlaceholder": "00100",
   "PayoutDetailsForm.postalCodeRequired": "This field is required",
+  "PayoutDetailsForm.stateLabel": "State",
+  "PayoutDetailsForm.statePlaceholder": "Enter your state",
+  "PayoutDetailsForm.stateRequired": "This field is required",
   "PayoutDetailsForm.streetAddressLabel": "Street address",
   "PayoutDetailsForm.streetAddressPlaceholder": "Enter your street address…",
   "PayoutDetailsForm.streetAddressRequired": "This field is required",
